Ai Logic
Created Shooting Logic for AI.
Created AI Projectile.
Added three collision layers for Player, Enemy and Bullet.
Bullet can damage player game object.

+++ UGJ100TheEnd/Assets/C# Scripts/Interfaces/IDamagable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    public void Damaged(float damage);
+    public void Damaged(int damage);
 }
